Split-out component-related types in `wasmparser`
This commit splits the `types.rs` in `wasmparser` into a `types.rs` and
`component_types.rs`. The goal here is to split all component-related
logic to a separate file instead of having everything bundled together.
This will make it easier to conditionally include component-related bits
in the future and additionally make each module more understandable by
focusing only on one piece of functionality.
